+++ src/public/cortex-compute-engine.d.ts
@@ -1,9 +1,0 @@
-declare module '@cortex-js/compute-engine' {
-  export type ComputeEngine = any;
-  export type BoxedExpression = any;
-  export type SemiBoxedExpression = any;
-}
-
-declare module '@cortex-js/compute-engine/dist/types/math-json/math-json-format' {
-  export type Expression = any;
-}