+++ src/public/mathlive-ssr.ts
@@ -1,5 +1,3 @@
-/// <reference path="cortex-compute-engine.d.ts" />
-
 /**
  * Server-side rendering exports.
  *